movies-recommender-system/requirements.txt
mysql-connector-python=9.4.0
sqlalchemy=2.0.43
bcrypt=4.3.0